ngram testing - appears to be functional

+++ src/Solver/Cpp/analysis/NGram.h
@@ -6,7 +6,7 @@
 
 class BiGram: public Analyser {
 protected:
-	std::unique_ptr<std::array<int, 1<<(2*5)>> frequencies;
+	std::unique_ptr<std::array<float, 1<<(2*5)>> frequencies;
 public:
 	BiGram(std::string);
 	float score(std::vector<int>);
@@ -14,7 +14,7 @@ public:
 
 class TriGram : public Analyser {
 protected:
-	std::unique_ptr<std::array<int, 1 << (3 * 5)>> frequencies;
+	std::unique_ptr<std::array<float, 1 << (3 * 5)>> frequencies;
 public:
 	TriGram(std::string);
 	float score(std::vector<int>);
@@ -22,7 +22,7 @@ public:
 
 class QuadGram : public Analyser {
 protected:
-	std::unique_ptr<std::array<int, 1 << (4 * 5)>> frequencies;
+	std::unique_ptr<std::array<float, 1 << (4 * 5)>> frequencies;
 public:
 	QuadGram(std::string);
 	float score(std::vector<int>);
